Recipe parsers refactoring
