+++ src/main/java/mc/recraftors/dumpster/parsers/recipes/BlastingJsonParser.java
@@ -1,10 +1,10 @@
-package mc.recraftors.dumpster.recipes;
+package mc.recraftors.dumpster.parsers.recipes;
 
 import net.minecraft.recipe.BlastingRecipe;
 import net.minecraft.recipe.Recipe;
 
 @TargetRecipeType(BlastingJsonParser.TYPE)
-public final class BlastingJsonParser extends AbstractCookingJsonParser {
+public class BlastingJsonParser extends AbstractCookingJsonParser {
     public static final String TYPE = "minecraft:blasting";
 
     public BlastingJsonParser() {
