+++ src/main/java/mc/recraftors/dumpster/parsers/recipes/SmokingJsonParser.java
@@ -1,10 +1,10 @@
-package mc.recraftors.dumpster.recipes;
+package mc.recraftors.dumpster.parsers.recipes;
 
 import net.minecraft.recipe.Recipe;
 import net.minecraft.recipe.SmokingRecipe;
 
 @TargetRecipeType(SmokingJsonParser.TYPE)
-public final class SmokingJsonParser extends AbstractCookingJsonParser {
+public class SmokingJsonParser extends AbstractCookingJsonParser {
     public static final String TYPE = "minecraft:smoking";
     public SmokingJsonParser() {
         super(TYPE);
